# Improve performance of LookupImpl.LookupAnalysis via pre-built index