`LookupImpl.LookupAnalysis` takes `O(classpath length)` to run. In project with long classpath, `LookupImpl.LookupAnalysis` is noticeable in profiling. (According to #593)

This PR introduces a `binaryClassNameToAnalysis` lookup table to reduce lookup time complexity.

For memory consumption of `binaryClassNameToAnalysis`, assuming there are 10000 binary class names in the worst case, each with 50 characters (100 bytes), the total key size is 10000 * 100 bytes = 1 MB, which is small.

Closes #593.