@eed3si9n 2023-11-06
Overall LGTM, but if would be cool if we could include either a benchmark or some sample project that demonstrates the improvement.