@Friendseeker 2023-11-06
Here's the two recent CI benchmark results with & without the optimization:

https://github.com/sbt/zinc/actions/runs/6773539163/job/18408627550 and https://github.com/sbt/zinc/actions/runs/6764410486/job/18382666918

Here's some excerpts:

Without optimization:
```
[info] Result "xsbt.HotScalacBenchmark.action":
[info]   N = 5
[info]   mean =  23098.871 ±(99.9%) 497.066 ms/op
[info]   Histogram, ms/op:
[info]     [22900.000, 22925.000) = 1 
[info]     [22925.000, 22950.000) = 0 
[info]     [22950.000, 22975.000) = 0 
[info]     [22975.000, 23000.000) = 0 
[info]     [23000.000, 23025.000) = 0 
[info]     [23025.000, 23050.000) = 0 
[info]     [23050.000, 23075.000) = 1 
[info]     [23075.000, 23100.000) = 1 
[info]     [23100.000, 23125.000) = 0 
[info]     [23125.000, 23150.000) = 0 
[info]     [23150.000, 23175.000) = 0 
[info]     [23175.000, 23200.000) = 1 
[info]     [23200.000, 23225.000) = 0 
[info]     [23225.000, 23250.000) = 0 
[info]     [23250.000, 23275.000) = 1 
[info]   Percentiles, ms/op:
[info]       p(0.0000) =  22917.677 ms/op
[info]      p(50.0000) =  23085.449 ms/op
[info]      p(90.0000) =  23253.221 ms/op
[info]      p(95.0000) =  23253.221 ms/op
[info]      p(99.0000) =  23253.221 ms/op
[info]      p(99.9000) =  23253.221 ms/op
[info]      p(99.9900) =  23253.221 ms/op
[info]      p(99.9990) =  23253.221 ms/op
[info]      p(99.9999) =  23253.221 ms/op
[info]     p(100.0000) =  23253.221 ms/op
[info] Secondary result "xsbt.HotScalacBenchmark.action:gc.alloc.rate":
[info]   130.140 ±(99.9%) 3.472 MB/sec [Average]
[info]   (min, avg, max) = (129.191, 130.140, 131.328), stdev = 0.902
[info]   CI (99.9%): [126.668, 133.613] (assumes normal distribution)
[info] Secondary result "xsbt.HotScalacBenchmark.action:gc.alloc.rate.norm":
[info]   3156755382.400 ±(99.9%) 9353105.171 B/op [Average]
[info]   (min, avg, max) = (3154084416.000, 3156755382.400, 3160329016.000), stdev = 2428971.711
[info]   CI (99.9%): [3147402277.229, 3166108487.571] (assumes normal distribution)
[info] Secondary result "xsbt.HotScalacBenchmark.action:gc.count":
[info]   90.000 ±(99.9%) 0.001 counts [Sum]
[info]   (min, avg, max) = (17.000, 18.000, 22.000), stdev = 2.236
[info]   CI (99.9%): [90.000, 90.000] (assumes normal distribution)
[info] Secondary result "xsbt.HotScalacBenchmark.action:gc.time":
[info]   7154.000 ±(99.9%) 0.001 ms [Sum]
[info]   (min, avg, max) = (1346.000, 1430.800, 1511.000), stdev = 64.325
[info]   CI (99.9%): [7154.000, 7154.000] (assumes normal distribution)
```

With optimization:
```
[info] Result "xsbt.HotScalacBenchmark.action":
[info]   N = 5
[info]   mean =  18656.264 ±(99.9%) 612.879 ms/op
[info]   Histogram, ms/op:
[info]     [18500.000, 18550.000) = 1 
[info]     [18550.000, 18600.000) = 1 
[info]     [18600.000, 18650.000) = 1 
[info]     [18650.000, 18700.000) = 1 
[info]     [18700.000, 18750.000) = 0 
[info]     [18750.000, 18800.000) = 0 
[info]     [18800.000, 18850.000) = 0 
[info]     [18850.000, 18900.000) = 0 
[info]     [18900.000, 18950.000) = 1 
[info]   Percentiles, ms/op:
[info]       p(0.0000) =  18522.046 ms/op
[info]      p(50.0000) =  18622.710 ms/op
[info]      p(90.0000) =  18924.700 ms/op
[info]      p(95.0000) =  18924.700 ms/op
[info]      p(99.0000) =  18924.700 ms/op
[info]      p(99.9000) =  18924.700 ms/op
[info]      p(99.9900) =  18924.700 ms/op
[info]      p(99.9990) =  18924.700 ms/op
[info]      p(99.9999) =  18924.700 ms/op
[info]     p(100.0000) =  18924.700 ms/op
[info] Secondary result "xsbt.HotScalacBenchmark.action:gc.alloc.rate":
[info]   160.312 ±(99.9%) 5.542 MB/sec [Average]
[info]   (min, avg, max) = (158.092, 160.312, 161.700), stdev = 1.439
[info]   CI (99.9%): [154.770, 165.854] (assumes normal distribution)
[info] Secondary result "xsbt.HotScalacBenchmark.action:gc.alloc.rate.norm":
[info]   3142377318.400 ±(99.9%) 4212317.767 B/op [Average]
[info]   (min, avg, max) = (3140712616.000, 3142377318.400, 3143361136.000), stdev = 1093925.547
[info]   CI (99.9%): [3138165000.633, 3146589636.167] (assumes normal distribution)
[info] Secondary result "xsbt.HotScalacBenchmark.action:gc.count":
[info]   104.000 ±(99.9%) 0.001 counts [Sum]
[info]   (min, avg, max) = (20.000, 20.800, 22.000), stdev = 1.095
[info]   CI (99.9%): [104.000, 104.000] (assumes normal distribution)
[info] Secondary result "xsbt.HotScalacBenchmark.action:gc.time":
[info]   6546.000 ±(99.9%) 0.001 ms [Sum]
[info]   (min, avg, max) = (1258.000, 1309.200, 1333.000), stdev = 31.815
[info]   CI (99.9%): [6546.000, 6546.000] (assumes normal distribution)
```

So the numbers do seem to be smaller. However I don't know if the CI uses the same machine each time. If CI benchmark results is not truth-worthy enough, I can do testing locally on my laptop.