@Friendseeker 2023-11-06
I had checked CI results for `xsbt.HotScalacBenchmark.action` across different recent PRs
```
1270: 23870.623 ±      897.960   ms/op

1272: 15522.280 ±      290.355   ms/op

1274:  22776.748 ±      824.291   ms/op

1274:  20984.942 ±     1091.008   ms/op
```

So... Cross comparison between different CI runs is meaningless. I shall run benchmarks locally instead to gather data.