@Friendseeker 2023-11-07
Here's the benchmark result:

With the optimization: 
```
[info] HotScalacBenchmark.action                         /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample   10        6817.422 ±      112.969   ms/op
[info] HotScalacBenchmark.action:gc.alloc.rate           /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample    5         440.283 ±       22.578  MB/sec
[info] HotScalacBenchmark.action:gc.alloc.rate.norm      /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample    5  3173548741.600 ± 12227671.111    B/op
[info] HotScalacBenchmark.action:gc.count                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample    5         214.000                 counts
[info] HotScalacBenchmark.action:gc.time                 /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample    5        3105.000                     ms
[info] HotScalacBenchmark.action:p0.00                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             6744.441                  ms/op
[info] HotScalacBenchmark.action:p0.50                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             6794.772                  ms/op
[info] HotScalacBenchmark.action:p0.90                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             6944.929                  ms/op
[info] HotScalacBenchmark.action:p0.95                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             6945.767                  ms/op
[info] HotScalacBenchmark.action:p0.99                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             6945.767                  ms/op
[info] HotScalacBenchmark.action:p0.999                  /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             6945.767                  ms/op
[info] HotScalacBenchmark.action:p0.9999                 /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             6945.767                  ms/op
[info] HotScalacBenchmark.action:p1.00                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             6945.767                  ms/op
[info] HotShapelessBenchmark.action                      /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample   18        3321.423 ±       27.155   ms/op
[info] HotShapelessBenchmark.action:gc.alloc.rate        /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample    5         463.219 ±       20.049  MB/sec
[info] HotShapelessBenchmark.action:gc.alloc.rate.norm   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample    5  1622412368.400 ±  5928171.244    B/op
[info] HotShapelessBenchmark.action:gc.count             /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample    5         213.000                 counts
[info] HotShapelessBenchmark.action:gc.time              /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample    5        3202.000                     ms
[info] HotShapelessBenchmark.action:p0.00                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             3271.557                  ms/op
[info] HotShapelessBenchmark.action:p0.50                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             3317.694                  ms/op
[info] HotShapelessBenchmark.action:p0.90                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             3364.671                  ms/op
[info] HotShapelessBenchmark.action:p0.95                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             3372.220                  ms/op
[info] HotShapelessBenchmark.action:p0.99                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             3372.220                  ms/op
[info] HotShapelessBenchmark.action:p0.999               /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             3372.220                  ms/op
[info] HotShapelessBenchmark.action:p0.9999              /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             3372.220                  ms/op
[info] HotShapelessBenchmark.action:p1.00                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true  sample             3372.220                  ms/op
[info] ColdScalacBenchmark.action                        /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true      ss    5       22372.120 ±      649.988   ms/op
[info] ColdScalacBenchmark.action:gc.alloc.rate          /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true      ss    5         150.684 ±        3.944  MB/sec
[info] ColdScalacBenchmark.action:gc.alloc.rate.norm     /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true      ss    5  3625515476.800 ± 16153258.595    B/op
[info] ColdScalacBenchmark.action:gc.count               /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true      ss    5         296.000                 counts
[info] ColdScalacBenchmark.action:gc.time                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true      ss    5        2373.000                     ms
[info] ColdShapelessBenchmark.action                     /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true      ss    5       16251.642 ±      673.525   ms/op
[info] ColdShapelessBenchmark.action:gc.alloc.rate       /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true      ss    5         114.146 ±        4.107  MB/sec
[info] ColdShapelessBenchmark.action:gc.alloc.rate.norm  /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true      ss    5  2003152908.800 ± 13875755.814    B/op
[info] ColdShapelessBenchmark.action:gc.count            /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true      ss    5         190.000                 counts
[info] ColdShapelessBenchmark.action:gc.time             /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_3a4aced9           true      ss    5        1626.000                     ms
[success] Total time: 586 s (09:46), completed Nov 6, 2023, 10:31:46 p.m.
```
Without optimization: 
```
[info] HotScalacBenchmark.action                         /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample   10        6767.929 ±       69.926   ms/op
[info] HotScalacBenchmark.action:gc.alloc.rate           /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample    5         442.517 ±       16.960  MB/sec
[info] HotScalacBenchmark.action:gc.alloc.rate.norm      /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample    5  3158922152.800 ±  7168994.553    B/op
[info] HotScalacBenchmark.action:gc.count                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample    5         208.000                 counts
[info] HotScalacBenchmark.action:gc.time                 /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample    5        2993.000                     ms
[info] HotScalacBenchmark.action:p0.00                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             6710.886                  ms/op
[info] HotScalacBenchmark.action:p0.50                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             6773.801                  ms/op
[info] HotScalacBenchmark.action:p0.90                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             6835.038                  ms/op
[info] HotScalacBenchmark.action:p0.95                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             6836.716                  ms/op
[info] HotScalacBenchmark.action:p0.99                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             6836.716                  ms/op
[info] HotScalacBenchmark.action:p0.999                  /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             6836.716                  ms/op
[info] HotScalacBenchmark.action:p0.9999                 /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             6836.716                  ms/op
[info] HotScalacBenchmark.action:p1.00                   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             6836.716                  ms/op
[info] HotShapelessBenchmark.action                      /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample   18        3321.889 ±       52.968   ms/op
[info] HotShapelessBenchmark.action:gc.alloc.rate        /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample    5         460.968 ±       26.362  MB/sec
[info] HotShapelessBenchmark.action:gc.alloc.rate.norm   /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample    5  1616603032.400 ±  4094565.358    B/op
[info] HotShapelessBenchmark.action:gc.count             /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample    5         208.000                 counts
[info] HotShapelessBenchmark.action:gc.time              /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample    5        3362.000                     ms
[info] HotShapelessBenchmark.action:p0.00                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             3250.586                  ms/op
[info] HotShapelessBenchmark.action:p0.50                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             3309.306                  ms/op
[info] HotShapelessBenchmark.action:p0.90                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             3428.005                  ms/op
[info] HotShapelessBenchmark.action:p0.95                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             3439.329                  ms/op
[info] HotShapelessBenchmark.action:p0.99                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             3439.329                  ms/op
[info] HotShapelessBenchmark.action:p0.999               /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             3439.329                  ms/op
[info] HotShapelessBenchmark.action:p0.9999              /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             3439.329                  ms/op
[info] HotShapelessBenchmark.action:p1.00                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true  sample             3439.329                  ms/op
[info] ColdScalacBenchmark.action                        /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true      ss    5       22429.638 ±      687.585   ms/op
[info] ColdScalacBenchmark.action:gc.alloc.rate          /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true      ss    5         150.384 ±        3.941  MB/sec
[info] ColdScalacBenchmark.action:gc.alloc.rate.norm     /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true      ss    5  3627138744.000 ± 31303175.452    B/op
[info] ColdScalacBenchmark.action:gc.count               /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true      ss    5         282.000                 counts
[info] ColdScalacBenchmark.action:gc.time                /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true      ss    5        2336.000                     ms
[info] ColdShapelessBenchmark.action                     /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true      ss    5       16221.570 ±      512.319   ms/op
[info] ColdShapelessBenchmark.action:gc.alloc.rate       /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true      ss    5         114.326 ±        1.850  MB/sec
[info] ColdShapelessBenchmark.action:gc.alloc.rate.norm  /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true      ss    5  2001709224.000 ± 16016911.514    B/op
[info] ColdShapelessBenchmark.action:gc.count            /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true      ss    5         194.000                 counts
[info] ColdShapelessBenchmark.action:gc.time             /var/folders/fm/f93y40ld5ql0h43m_7p6kkqh0000gn/T/sbt_89158142           true      ss    5        1616.000                     ms
[success] Total time: 585 s (09:45), completed Nov 6, 2023, 10:44:17 p.m.
```

The changes are within the error margin, so it seems this was a case of premature optimization. I suggest we discard this change to avoid unnecessary complexity in the codebase. Thank you for the opportunity for me to learn and improve through this process.